ingest_data/requirements.txt
# Storage & Database
chromadb>=0.4.24,<0.6.0
boto3>=1.34.51,<1.35.0

pandas==2.0.3
numpy>=1.26.2,<2.0.0
pyarrow>=17.0.0

langchain==0.3.26
langchain-core==0.3.70
langchain-community==0.3.26
sentence-transformers==2.6.1
huggingface-hub>=0.23 
langchain-chroma==0.1.4

wget==3.2
pydantic>=2.10.0